Updates and minor corrections. Unsuccessful validation.
Even after completing all the revisions, I was getting duplicate element definition error messages. Also, note the comments about how element definitions got moved from commonElementsMod to topicMod. Given the emphasis on RNG over XSD and my lack of expertise with XSD, I am referring this issue to WEK and moving on to the RNG section of this tutorial. If there is time and need later, I can come back to it.